# perf: improve Buffer.from(buf, 'base64') performance