Improve `Buffer.from(buffer, "base64")` performance by ~29x

```
$ deno run -A bench.mjs
7987.390458 ms

$ node bench.mjs
773.420042 ms

$ bun bench.mjs
737.7943750 ms

$ target/release/deno run -A bench.mjs
275.3805 ms
```

```js
import { Buffer } from "node:buffer";

const buffer = Buffer.alloc(1024 * 10, "latin1");

const start = performance.now();
for (let i = 0; i < 1000000; i++) {
  Buffer.from(buffer, "base64");
}
console.log(performance.now() - start);
```